fix(dropdown-menu,context-menu,menubar): improve brn data attributes for menus
Extend the BrnMenuDirective class to set the side and state
attributes on the menu host. This update introduces new
methods to determine the correct side where a submenu will
appear and consequently set the attribute to the host. It
also handles the menu's state when the host is about to
close. A workaround was applied to approximate the
appropriate submenu side.